+++ bmtrain/csrc/cuda/cross_entropy.cu
@@ -1,5 +1,4 @@
-// #include <cuda_fp16.h>
-#include <cuda_bf16.h>
+#include <cuda_fp16.h>
 #include <torch/extension.h>
 #include <ATen/cuda/CUDAContext.h>
 #include "reduce.cuh"
@@ -8,9 +7,9 @@ namespace {
 // blocks <m>,      threads<1024>
 __global__ void cross_entropy_forward(
     int64_t n,
-    const nv_bfloat16 *input,      // (m, n)
+    const half *input,      // (m, n)
     const int32_t *target,  // (m)
-    nv_bfloat16 *softmax,          // (m, n)
+    half *softmax,          // (m, n)
     float *output,          // (m)
     int32_t ignore_index
 ) {
@@ -18,24 +17,24 @@ __global__ void cross_entropy_forward(
 
     float local_max = -INFINITY;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_max = fmaxf(__bfloat162float(input[base_idx + i]), local_max);
+        local_max = fmaxf(__half2float(input[base_idx + i]), local_max);
     }
 
     local_max = fmaxf(block_allreduce_max(local_max), -1e6);
     
     float local_sum = 0;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_sum += expf(__bfloat162float(input[base_idx + i]) - local_max);
+        local_sum += expf(__half2float(input[base_idx + i]) - local_max);
     }
     local_sum = block_allreduce_sum(local_sum) + 1e-10; // avoid nan
     
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        softmax[base_idx + i] = __float2bfloat16( expf(__bfloat162float(input[base_idx + i]) - local_max) / local_sum );
+        softmax[base_idx + i] = __float2half( expf(__half2float(input[base_idx + i]) - local_max) / local_sum );
     }
 
     if (threadIdx.x == 0) {
         if (target[blockIdx.x] != ignore_index) {
-            output[blockIdx.x] = -__bfloat162float(input[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
+            output[blockIdx.x] = -__half2float(input[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
         } else {
             output[blockIdx.x] = 0;
         }
@@ -47,21 +46,21 @@ __global__ void cross_entropy_backward(
     int64_t n,
     const float *grad_output,   // (m)
     const int32_t *target,      // (m)
-    const nv_bfloat16 *softmax,        // (m, n)
-    nv_bfloat16 *grad_input,           // (m, n)
+    const half *softmax,        // (m, n)
+    half *grad_input,           // (m, n)
     int32_t ignore_index
 ) {
     int64_t base_idx = blockIdx.x * n;
 
     int32_t t = target[blockIdx.x];
     if (t == ignore_index) {
-        nv_bfloat16 v = __float2bfloat16(0.);
+        half v = __float2half(0.);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             grad_input[base_idx + i] = v;
         }
     }
     else {
-        nv_bfloat16 v = __float2bfloat16(grad_output[blockIdx.x]);
+        half v = __float2half(grad_output[blockIdx.x]);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             grad_input[base_idx + i] = i==t ? __hsub(__hmul(softmax[base_idx + i], v), v) : __hmul(softmax[base_idx + i], v);
         }
@@ -71,7 +70,7 @@ __global__ void cross_entropy_backward(
 // blocks <m>,      threads<1024>
 __global__ void cross_entropy_forward_inplace(
     int64_t n,
-    nv_bfloat16 *x,                // (m, n)
+    half *x,                // (m, n)
     const int32_t *target,  // (m)
     float *output,          // (m)
     int32_t ignore_index
@@ -80,19 +79,19 @@ __global__ void cross_entropy_forward_inplace(
 
     float local_max = -INFINITY;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_max = fmaxf(__bfloat162float(x[base_idx + i]), local_max);
+        local_max = fmaxf(__half2float(x[base_idx + i]), local_max);
     }
     local_max = fmaxf(block_allreduce_max(local_max), -1e6);
     
     float local_sum = 0;
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        local_sum += expf(__bfloat162float(x[base_idx + i]) - local_max);
+        local_sum += expf(__half2float(x[base_idx + i]) - local_max);
     }
     local_sum = block_allreduce_sum(local_sum) + 1e-10; // avoid nan
 
     if (threadIdx.x == 0) {
         if (target[blockIdx.x] != ignore_index) {
-            output[blockIdx.x] = -__bfloat162float(x[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
+            output[blockIdx.x] = -__half2float(x[base_idx + target[blockIdx.x]]) + local_max + logf(local_sum);
         } else {
             output[blockIdx.x] = 0;
         }
@@ -101,7 +100,7 @@ __global__ void cross_entropy_forward_inplace(
     __syncthreads();
     
     for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
-        x[base_idx + i] = __float2bfloat16( expf(__bfloat162float(x[base_idx + i]) - local_max) / local_sum );
+        x[base_idx + i] = __float2half( expf(__half2float(x[base_idx + i]) - local_max) / local_sum );
     }
 }
 
@@ -110,20 +109,20 @@ __global__ void cross_entropy_backward_inplace(
     int64_t n,
     const float *grad_output,   // (m)
     const int32_t *target,      // (m)
-    nv_bfloat16 *x,                    // (m, n)
+    half *x,                    // (m, n)
     int32_t ignore_index
 ) {
     int64_t base_idx = blockIdx.x * n;
 
     int32_t t = target[blockIdx.x];
     if (t == ignore_index) {
-        nv_bfloat16 v = __float2bfloat16(0.);
+        half v = __float2half(0.);
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             x[base_idx + i] = v;
         }
     }
     else {
-        nv_bfloat16 v = __float2bfloat16(grad_output[blockIdx.x]);
+        half v = __float2half(grad_output[blockIdx.x]);
         __syncthreads();
         for (int64_t i = threadIdx.x; i < n; i += blockDim.x) {
             x[base_idx + i] = i==t ? __hsub(__hmul(x[base_idx + i], v), v) : __hmul(x[base_idx + i], v);
@@ -141,9 +140,9 @@ void cross_entropy_forward_launcher(
     torch::Tensor &output,
     int32_t ignore_index
 ) {
-    auto input_ptr = reinterpret_cast<nv_bfloat16*>(input.data_ptr<at::BFloat16>());
+    auto input_ptr = reinterpret_cast<half*>(input.data_ptr<at::Half>());
     auto target_ptr = target.data_ptr<int32_t>();
-    auto softmax_ptr = reinterpret_cast<nv_bfloat16*>(softmax.data_ptr<at::BFloat16>());
+    auto softmax_ptr = reinterpret_cast<half*>(softmax.data_ptr<at::Half>());
     auto output_ptr = output.data_ptr<float>();
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
@@ -160,8 +159,8 @@ void cross_entropy_backward_launcher(
 ) {
     auto output_ptr = grad_output.data_ptr<float>();
     auto target_ptr = target.data_ptr<int32_t>();
-    auto softmax_ptr = reinterpret_cast<nv_bfloat16*>(softmax.data_ptr<at::BFloat16>());
-    auto input_ptr = reinterpret_cast<nv_bfloat16*>(grad_input.data_ptr<at::BFloat16>());
+    auto softmax_ptr = reinterpret_cast<half*>(softmax.data_ptr<at::Half>());
+    auto input_ptr = reinterpret_cast<half*>(grad_input.data_ptr<at::Half>());
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
     cross_entropy_backward<<<m, threads, 0, stream.stream()>>>(n, output_ptr, target_ptr, softmax_ptr, input_ptr, ignore_index);
@@ -174,7 +173,7 @@ void cross_entropy_forward_inplace_launcher(
     torch::Tensor &output,
     int32_t ignore_index
 ) {
-    auto x_ptr = reinterpret_cast<nv_bfloat16*>(x.data_ptr<at::BFloat16>());
+    auto x_ptr = reinterpret_cast<half*>(x.data_ptr<at::Half>());
     auto target_ptr = target.data_ptr<int32_t>();
     auto output_ptr = output.data_ptr<float>();
     int32_t threads = 1024;
@@ -191,7 +190,7 @@ void cross_entropy_backward_inplace_launcher(
 ) {
     auto output_ptr = grad_output.data_ptr<float>();
     auto target_ptr = target.data_ptr<int32_t>();
-    auto x_ptr = reinterpret_cast<nv_bfloat16*>(x.data_ptr<at::BFloat16>());
+    auto x_ptr = reinterpret_cast<half*>(x.data_ptr<at::Half>());
     int32_t threads = 1024;
     auto stream = at::cuda::getCurrentCUDAStream();
     cross_entropy_backward_inplace<<<m, threads, 0, stream.stream()>>>(n, output_ptr, target_ptr, x_ptr, ignore_index);
